+++ mobile/android/android-components/components/feature/sitepermissions/src/test/java/mozilla/components/feature/sitepermissions/SitePermissionsRulesTest.kt
@@ -59,7 +59,8 @@ class SitePermissionsRulesTest {
             microphone = BLOCKED,
             autoplayAudible = ASK_TO_ALLOW,
             autoplayInaudible = BLOCKED,
-            persistentStorage = BLOCKED
+            persistentStorage = BLOCKED,
+            mediaKeySystemAccess = ASK_TO_ALLOW
         )
 
         val mockRequest: PermissionRequest = mock()
@@ -95,6 +96,10 @@ class SitePermissionsRulesTest {
         doReturn(listOf(Permission.ContentPersistentStorage())).`when`(mockRequest).permissions
         action = rules.getActionFrom(mockRequest)
         assertEquals(action, rules.persistentStorage)
+
+        doReturn(listOf(Permission.ContentMediaKeySystemAccess())).`when`(mockRequest).permissions
+        action = rules.getActionFrom(mockRequest)
+        assertEquals(action, rules.mediaKeySystemAccess)
     }
 
     @Test
@@ -106,7 +111,8 @@ class SitePermissionsRulesTest {
             notification = ASK_TO_ALLOW,
             microphone = BLOCKED,
             autoplayInaudible = ASK_TO_ALLOW,
-            autoplayAudible = ASK_TO_ALLOW
+            autoplayAudible = ASK_TO_ALLOW,
+            mediaKeySystemAccess = ASK_TO_ALLOW
         )
 
         val mockRequest: PermissionRequest = mock()
@@ -122,7 +128,8 @@ class SitePermissionsRulesTest {
             microphone = ASK_TO_ALLOW,
             autoplayInaudible = BLOCKED,
             autoplayAudible = BLOCKED,
-            persistentStorage = BLOCKED
+            persistentStorage = BLOCKED,
+            mediaKeySystemAccess = ASK_TO_ALLOW
         )
 
         action = rules.getActionFrom(mockRequest)
@@ -140,6 +147,7 @@ class SitePermissionsRulesTest {
                 microphone = Status.BLOCKED,
                 autoplayInaudible = Status.NO_DECISION,
                 autoplayAudible = Status.NO_DECISION,
+                mediaKeySystemAccess = Status.BLOCKED,
                 savedAt = 1L
         )
 
@@ -150,7 +158,8 @@ class SitePermissionsRulesTest {
                 microphone = BLOCKED,
                 autoplayInaudible = ASK_TO_ALLOW,
                 autoplayAudible = ASK_TO_ALLOW,
-                persistentStorage = BLOCKED
+                persistentStorage = BLOCKED,
+                mediaKeySystemAccess = BLOCKED
         )
 
         val convertedSitePermissions = rules.toSitePermissions(origin = "origin", savedAt = 1L)
@@ -163,6 +172,7 @@ class SitePermissionsRulesTest {
         assertEquals(expectedSitePermission.autoplayInaudible, convertedSitePermissions.autoplayInaudible)
         assertEquals(expectedSitePermission.autoplayAudible, convertedSitePermissions.autoplayAudible)
         assertEquals(expectedSitePermission.localStorage, convertedSitePermissions.localStorage)
+        assertEquals(expectedSitePermission.mediaKeySystemAccess, convertedSitePermissions.mediaKeySystemAccess)
         assertEquals(expectedSitePermission.savedAt, convertedSitePermissions.savedAt)
     }
 }